Refactored root endpoint and added service hook for things that should happen after exploitation is done
